requirements.txt
requests==2.28.1             
psycopg2-binary==2.9.6      
sqlalchemy==1.4.46          
numpy>=1.21.6,<1.22.0        
pandas>=1.3.5,<1.4.0         
scikit-learn>=0.24.2,<0.25.0 
huggingface_hub==0.8.1       
sentence-transformers==2.2.0 
nltk>=3.6.3,<3.7.0           
pgvector==0.1.0              
pytest>=6.2.5,<7.0.0      